Refatoração da classe PessoaController
