--- AppClienteVIP/app/src/main/java/br/com/dnsistemas/appclientevip/controller/PessoaContoller.java
@@ -1,9 +1,0 @@
-package br.com.dnsistemas.appclientevip.controller;
-
-import br.com.dnsistemas.appclientevip.model.Pessoa;
-
-public class PessoaContoller {
-    public void salvar(Pessoa pessoa) {
-
-    }
-}
