+++ AppClienteVIP/app/src/main/java/br/com/dnsistemas/appclientevip/controller/PessoaController.java
@@ -1,0 +1,40 @@
+package br.com.dnsistemas.appclientevip.controller;
+
+import android.content.SharedPreferences;
+
+import br.com.dnsistemas.appclientevip.model.Pessoa;
+import br.com.dnsistemas.appclientevip.view.MainActivity;
+
+public class PessoaController {
+
+    SharedPreferences preferences;
+    SharedPreferences.Editor listaVip;
+    public static final String NOME_PREFERENCES = "pref_listavip";
+
+    public PessoaController(MainActivity mainActivity) {
+        preferences = mainActivity.getSharedPreferences(NOME_PREFERENCES, 0);
+        listaVip = preferences.edit();
+    }
+
+    public void salvar(Pessoa pessoa) {
+        listaVip.putString("primeiroNome", pessoa.getPrimeiroNome());
+        listaVip.putString("sobrenome", pessoa.getSobrenome());
+        listaVip.putString("cursoDesejado", pessoa.getCursoDesejado());
+        listaVip.putString("telefoneContato", pessoa.getTelefoneContato());
+        listaVip.apply();
+    }
+
+    public Pessoa buscar(Pessoa pessoa) {
+        pessoa.setPrimeiroNome(preferences.getString("primeiroNome", "NA"));
+        pessoa.setSobrenome(preferences.getString("sobrenome", "NA"));
+        pessoa.setCursoDesejado(preferences.getString("cursoDesejado", "NA"));
+        pessoa.setTelefoneContato(preferences.getString("telefoneContato", "NA"));
+
+        return pessoa;
+    }
+
+    public void limpar() {
+        listaVip.clear();
+        listaVip.apply();
+    }
+}
